test update

+++ scripts/build_scripts/01_prepare_project.sh
@@ -2,20 +2,21 @@
 
 if [ -z "$IONIC_APP_VERSION_NUMBER" ]
   then
-    echo "${GREEN}No version parameter second argument given so using ${IONIC_APP_VERSION_NUMBER} as default version number...${NC}"
+    echo "${GREEN}Please provide IONIC_APP_VERSION_NUMBER...${NC}"
+    exit 1
 else
     echo -e "IONIC_APP_VERSION_NUMBER is $IONIC_APP_VERSION_NUMBER...${NC}"
 fi
 
 if [ -z "$INTERMEDIATE_PATH" ]
     then
-      $INTERMEDIATE_PATH="$PWD"
+      export INTERMEDIATE_PATH="$PWD"
       echo "No INTERMEDIATE_PATH given. Using $INTERMEDIATE_PATH..."
 fi
 
 if [ -z "$BUILD_PATH" ]
     then
-      $BUILD_PATH="$IONIC_PATH"/build
+      export BUILD_PATH="$IONIC_PATH"/build
       echo "No BUILD_PATH given. Using $BUILD_PATH..."
 fi
 
@@ -51,7 +52,16 @@ fi
 export LC_CTYPE=C
 export LANG=C
 echo -e "${GREEN}Replacing IONIC_APP_VERSION_NUMBER with ${IONIC_APP_VERSION_NUMBER}...${NC}"
-cd "${INTERMEDIATE_PATH}/apps" && find . -type f -exec sed -i '' -e 's/IONIC_APP_VERSION_NUMBER/'${IONIC_APP_VERSION_NUMBER}'/g' {} \; >> /dev/null 2>&1
+cp ${INTERMEDIATE_PATH}/config-template.xml ${INTERMEDIATE_PATH}/apps/${LOWERCASE_APP_NAME}/config.xml
+cd ${INTERMEDIATE_PATH}/apps/${LOWERCASE_APP_NAME}
+find . -type f -exec sed -i '' -e 's/IONIC_IOS_APP_VERSION_NUMBER/'${IONIC_IOS_APP_VERSION_NUMBER}'/g' {} \; >> /dev/null 2>&1
+find . -type f -exec sed -i '' -e 's/IONIC_APP_VERSION_NUMBER/'${IONIC_APP_VERSION_NUMBER}'/g' {} \; >> /dev/null 2>&1
+find . -type f -exec sed -i '' -e 's/APP_DISPLAY_NAME/'${APP_DISPLAY_NAME}'/g' {} \; >> /dev/null 2>&1
+find . -type f -exec sed -i '' -e 's/APP_IDENTIFIER/'${APP_IDENTIFIER}'/g' {} \; >> /dev/null 2>&1
+
+echo "MAKE SURE NOT TO USE QUOTES OR SPECIAL CHARACTERS WITH export APP_DESCRIPTION OR IT WILL NOT REPLACE PROPERLY"
+find . -type f -exec sed -i '' -e 's/APP_DESCRIPTION/'${APP_DESCRIPTION}'/g' {} \; >> /dev/null 2>&1
+
 export LANG=en_US.UTF-8
 
 echo -e "${GREEN}Copy ${LOWERCASE_APP_NAME} config and resource files${NC}"
@@ -64,7 +74,10 @@ cd "${INTERMEDIATE_PATH}"
 echo "Copying generated images from ${INTERMEDIATE_PATH}/resources/android to ${INTERMEDIATE_PATH}/www/img/"
 cp -R ${INTERMEDIATE_PATH}/resources/android/*  "${INTERMEDIATE_PATH}/www/img/"
 
+echo "Removing ${BUILD_PATH}/${LOWERCASE_APP_NAME}"
 rm -rf "${BUILD_PATH}/${LOWERCASE_APP_NAME}"
 
-echo -e "${GREEN}Copy ${APP_PRIVATE_CONFIG_PATH}/${LOWERCASE_APP_NAME}.config.js private config to ${INTERMEDIATE_PATH}/www/private_configs/${NC}"
-cp "${APP_PRIVATE_CONFIG_PATH}/${LOWERCASE_APP_NAME}.config.js" "${INTERMEDIATE_PATH}/www/private_configs/"
+if [ ! -f ${INTERMEDIATE_PATH}/www/private_configs//${LOWERCASE_APP_NAME}.config.js ]; then
+    echo -e "${GREEN}Copy ${APP_PRIVATE_CONFIG_PATH}/${LOWERCASE_APP_NAME}.config.js private config to ${INTERMEDIATE_PATH}/www/private_configs/${NC}"
+    cp "${APP_PRIVATE_CONFIG_PATH}/${LOWERCASE_APP_NAME}.config.js" "${INTERMEDIATE_PATH}/www/private_configs/"
+fi
